requirements.txt
pyyaml>=6.0
anthropic>=0.40.0
keyring>=25.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
openai-whisper>=20231117
